% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{apply_label_mapping}
\alias{apply_label_mapping}
\title{Apply an integer-taxon label mapping (label_mapping)
to an integer-based newick (where leaves are integers)
and produce a mapped Newick (where leaves are strings (taxa))}
\usage{
apply_label_mapping(newick, label_mapping_list)
}
\arguments{
\item{newick}{Newick representation of a tree with integer labels}

\item{label_mapping_list}{A vector where the index corresponds to the integer label}
}
\value{
Newick with string labels
}
\description{
For more details, see \code{create_label_mapping}.
}
\details{
Note 1: this does not check for the validity of the Newick string.

Note 2: the parent nodes are removed from the output,
but the branch lengths/annotations are kept.
}
