% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{b2}
\alias{b2}
\title{Compute the B2 index of a tree from Shao and Sokal (1990).}
\usage{
b2(vector)
}
\arguments{
\item{vector}{phylo2vec vector representation of a tree topology}
}
\value{
B2 index (numeric)
}
\description{
The B2 index is a measure of tree balance based on the probabilities of random walks from the root to each leaf.
For a binary tree, the B2 index can be calculated as the sum of the depth of each leaf multiplied by 2 raised to the power of negative depth of that leaf.
Higher values indicate more balanced trees, while lower values indicate more imbalanced trees.
For more details, see https://doi.org/10.1007/s00285-021-01662-7.
}
