% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{check_path}
\alias{check_path}
\title{Check if the file path has an accepted extension.}
\usage{
check_path(filepath, filetype)
}
\arguments{
\item{filepath}{File path to check.}

\item{filetype}{Type of file: "array" or "newick".}
}
\description{
Raises an error if the file extension is not supported
}
