% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{check_v}
\alias{check_v}
\title{Validate a phylo2vec vector}
\usage{
check_v(vector)
}
\arguments{
\item{vector}{phylo2vec vector representation of a tree topology}
}
\description{
Raises an error if the vector is invalid.
}
