% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{cophenetic_distances}
\alias{cophenetic_distances}
\title{Get the topological cophenetic distance matrix of a Phylo2Vec vector}
\usage{
cophenetic_distances(vector)
}
\description{
Get the topological cophenetic distance matrix of a Phylo2Vec vector
}
