% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{cophenetic_from_matrix}
\alias{cophenetic_from_matrix}
\title{Get the cophenetic distance matrix of a Phylo2Vec matrix}
\usage{
cophenetic_from_matrix(matrix, unrooted)
}
\description{
Get the cophenetic distance matrix of a Phylo2Vec matrix
}
