% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{cophenetic_from_vector}
\alias{cophenetic_from_vector}
\title{Get the topological cophenetic distance matrix of a Phylo2Vec vector}
\usage{
cophenetic_from_vector(vector)
}
\description{
Get the topological cophenetic distance matrix of a Phylo2Vec vector
}
