% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{cov}
\alias{cov}
\title{Compute the variance-covariance matrix of a Phylo2Vec
vector (topological) or matrix (from branch lengths).}
\usage{
cov(vector_or_matrix)
}
\arguments{
\item{vector_or_matrix}{Phylo2Vec vector (ndim == 1)/matrix (ndim == 2)}
}
\value{
Variance-covariance matrix
}
\description{
Compute the variance-covariance matrix of a Phylo2Vec
vector (topological) or matrix (from branch lengths).
}
