% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{create_label_mapping}
\alias{create_label_mapping}
\title{Create an integer-taxon label mapping (label_mapping)
from a string-based newick (where leaves are strings)
and produce a mapped integer-based newick (where leaves are integers).}
\usage{
create_label_mapping(newick)
}
\description{
Note 1: this does not check for the validity of the Newick string.
}
\details{
Note 2: the parent nodes are removed from the output,
but the branch lengths/annotations are kept.
}
