% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{find_num_leaves}
\alias{find_num_leaves}
\title{Find the number of leaves in a Newick string}
\usage{
find_num_leaves(newick)
}
\arguments{
\item{newick}{Newick representation of a tree}
}
\value{
Number of leaves
}
\description{
Find the number of leaves in a Newick string
}
