% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_newick.R
\name{from_newick}
\alias{from_newick}
\title{Convert a Newick string to a phylo2vec vector or matrix}
\usage{
from_newick(newick)
}
\arguments{
\item{newick}{Newick string representation of the tree}
}
\value{
A phylo2vec matrix (tree with branch lengths)
or a phylo2vec vector (tree topology only)
}
\description{
Convert a Newick string to a phylo2vec vector or matrix
}
