% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{from_pairs}
\alias{from_pairs}
\title{Convert a pairs matrix to a phylo2vec vector}
\usage{
from_pairs(pairs)
}
\arguments{
\item{pairs}{Pairs representation (shape: (n_leaves - 1, 2))}
}
\value{
phylo2vec vector representation
}
\description{
Convert a pairs matrix to a phylo2vec vector
}
