% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_node_depth}
\alias{get_node_depth}
\title{Get the depth of a node in a phylo2vec tree
Depth is the distance from root to that node. Root has depth 0.
For vectors, this is the topological depth. For matrices, this is the depth with branch lengths.}
\usage{
get_node_depth(tree, node)
}
\arguments{
\item{tree}{A phylo2vec tree}

\item{node}{The node to get the depth of (0-indexed)}
}
\value{
The depth of the node
}
\description{
Get the depth of a node in a phylo2vec tree
Depth is the distance from root to that node. Root has depth 0.
For vectors, this is the topological depth. For matrices, this is the depth with branch lengths.
}
