% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_node_depths}
\alias{get_node_depths}
\title{Get the depths of all nodes in a phylo2vec tree
Depth is the distance from root to each node. Root has depth 0.}
\usage{
get_node_depths(tree)
}
\arguments{
\item{tree}{A phylo2vec tree}
}
\value{
A vector of depths for all nodes (length: 2*n_leaves - 1)
}
\description{
Get the depths of all nodes in a phylo2vec tree
Depth is the distance from root to each node. Root has depth 0.
}
