% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{has_branch_lengths}
\alias{has_branch_lengths}
\title{Check if a newick string has branch lengths}
\usage{
has_branch_lengths(newick)
}
\arguments{
\item{newick}{Newick representation of a tree}
}
\value{
TRUE if the newick has branch lengths, FALSE otherwise
}
\description{
Check if a newick string has branch lengths
}
