% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{incidence_coo}
\alias{incidence_coo}
\title{Get the incidence matrix of a Phylo2Vec vector in COUP format}
\usage{
incidence_coo(input_vector)
}
\description{
Get the incidence matrix of a Phylo2Vec vector in COUP format
}
