% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{incidence_csr}
\alias{incidence_csr}
\title{Get the incidence matrix of a Phylo2Vec vector in CSR format}
\usage{
incidence_csr(input_vector)
}
\description{
Get the incidence matrix of a Phylo2Vec vector in CSR format
}
