% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{incidence_dense}
\alias{incidence_dense}
\title{Get the incidence matrix of a Phylo2Vec vector in dense format}
\usage{
incidence_dense(input_vector)
}
\description{
Get the incidence matrix of a Phylo2Vec vector in dense format
}
