% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{leaf_depth_variance}
\alias{leaf_depth_variance}
\title{Compute the variance of leaf depths in a tree.}
\usage{
leaf_depth_variance(vector)
}
\arguments{
\item{vector}{phylo2vec vector representation of a tree topology}
}
\value{
Variance of leaf depths (numeric)
}
\description{
Higher values indicate more imbalanced trees, while lower values indicate more balanced trees.
}
