% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_newick}
\alias{load_newick}
\title{Read a newick file into a Phylo2Vec vector or matrix.}
\usage{
load_newick(filepath_or_buffer)
}
\arguments{
\item{filepath_or_buffer}{File path or string containing
a Newick-formatted tree}
}
\value{
A vector (ndim == 1) or matrix (ndim == 2)
which satisfies Phylo2Vec constraints
}
\description{
Read a newick file into a Phylo2Vec vector or matrix.
}
