% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_p2v}
\alias{load_p2v}
\title{Read a text/csv file into a Phylo2Vec vector or matrix.}
\usage{
load_p2v(filepath, delimiter = ",")
}
\arguments{
\item{filepath}{File path to read.}

\item{delimiter}{Character used to separate values, defaults to ",".}
}
\value{
A vector (ndim == 1) or matrix (ndim == 2).
which satisfies Phylo2Vec constraints
}
\description{
Read a text/csv file into a Phylo2Vec vector or matrix.
}
