% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{pre_precision}
\alias{pre_precision}
\title{Get a precursor of the precision matrix of a phylo2vec tree
The precision matrix is the inverse of the variance-covariance matrix.
The precision matrix can be obtained using Schur's complement on this precursor.
Output shape: (2 * (n_leaves - 1), 2 * (n_leaves- 1)]}
\usage{
pre_precision(tree)
}
\arguments{
\item{tree}{A phylo2vec tree}
}
\value{
A precursor of the precision matrix (shape: (2 * (n_leaves - 1), 2 * (n_leaves- 1)))
}
\description{
Get a precursor of the precision matrix of a phylo2vec tree
The precision matrix is the inverse of the variance-covariance matrix.
The precision matrix can be obtained using Schur's complement on this precursor.
Output shape: (2 * (n_leaves - 1), 2 * (n_leaves- 1)]
}
