% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{pre_precision_from_matrix}
\alias{pre_precision_from_matrix}
\title{Get the precision matrix of a Phylo2Vec matrix}
\usage{
pre_precision_from_matrix(matrix)
}
\description{
Get the precision matrix of a Phylo2Vec matrix
}
