% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{pre_precision_from_vector}
\alias{pre_precision_from_vector}
\title{Get the precision matrix of a Phylo2Vec vector}
\usage{
pre_precision_from_vector(vector)
}
\description{
Get the precision matrix of a Phylo2Vec vector
}
