% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{queue_shuffle}
\alias{queue_shuffle}
\title{Produce an ordered version (i.e., birth-death process version)
of a Phylo2Vec vector using the Queue Shuffle algorithm.}
\usage{
queue_shuffle(vector, shuffle_cherries)
}
\arguments{
\item{vector}{A Phylo2Vec vector (i.e., a vector of integers)}

\item{shuffle_cherries}{If true, the algorithm will shuffle cherries (i.e., pairs of leaves)}
}
\value{
A list with two elements:
\itemize{
\item \code{v}: The ordered Phylo2Vec vector
\item \code{mapping}: A mapping of the original labels to the new labels
}
}
\description{
Queue Shuffle ensures that the output tree is ordered,
while also ensuring a smooth path through the space of orderings
}
\details{
for more details, see https://doi.org/10.1093/gbe/evad213
}
