% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{remove_parent_labels}
\alias{remove_parent_labels}
\title{Remove parent labels from the Newick string}
\usage{
remove_parent_labels(newick)
}
\arguments{
\item{newick}{Newick representation of a tree with parent labels}
}
\value{
Newick string without parent labels
}
\description{
Remove parent labels from the Newick string
}
