% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{sackin}
\alias{sackin}
\title{Compute the Sackin index of a tree.}
\usage{
sackin(vector)
}
\arguments{
\item{vector}{phylo2vec vector representation of a tree topology}
}
\value{
Sackin index (numeric)
}
\description{
The Sackin index is a measure of tree imbalance, defined as the sum of the depths of all leaves in the tree.
Higher values indicate more imbalanced trees, while lower values indicate more balanced trees.
}
