% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{sample_matrix}
\alias{sample_matrix}
\title{Sample a random tree with branch lengths via Phylo2Vec}
\usage{
sample_matrix(n_leaves, ordered)
}
\description{
Sample a random tree with branch lengths via Phylo2Vec
}
