% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_tree}
\alias{sample_tree}
\title{Sample a random tree as a Phylo2Vec object}
\usage{
sample_tree(n_leaves, topology_only = FALSE, ordered = FALSE)
}
\arguments{
\item{n_leaves}{Number of leaves in the tree}

\item{topology_only}{If TRUE, a vector corresponding to the topology
is returned. IF FALSE, a matrix is returned}

\item{ordered}{If TRUE, sample an ordered tree}
}
\value{
A Phylo2Vec object (\code{integer} if topology_only, else \code{matrix})
}
\description{
Sample a random tree as a Phylo2Vec object
}
