% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{save_newick}
\alias{save_newick}
\title{Save a Phylo2Vec vector or matrix to Newick format into a file.}
\usage{
save_newick(vector_or_matrix, filepath, labels = NULL)
}
\arguments{
\item{vector_or_matrix}{A vector (ndim == 1) or matrix (ndim == 2)}

\item{filepath}{File path to save the Newick-formatted tree}

\item{labels}{Optional labels for the leaves of the tree}
}
\description{
Save a Phylo2Vec vector or matrix to Newick format into a file.
}
