% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{save_p2v}
\alias{save_p2v}
\title{Save a Phylo2Vec vector or matrix to a file.}
\usage{
save_p2v(vector_or_matrix, filepath, delimiter = ",")
}
\arguments{
\item{vector_or_matrix}{A vector (ndim == 1) or matrix (ndim == 2)
which satisfies Phylo2Vec constraints}

\item{filepath}{Path to the output file}

\item{delimiter}{Character used to separate values, defaults to ",".}
}
\description{
Save a Phylo2Vec vector or matrix to a file.
}
