% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_newick.R
\name{to_newick}
\alias{to_newick}
\title{Convert a Phylo2Vec vector or matrix to Newick format}
\usage{
to_newick(vector_or_matrix)
}
\arguments{
\item{vector_or_matrix}{Phylo2Vec matrix if branch lengths
are present, otherwise a vector}
}
\value{
Newick string representation of the tree
}
\description{
Convert a Phylo2Vec vector or matrix to Newick format
}
