% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{to_newick_from_matrix}
\alias{to_newick_from_matrix}
\title{Recover a rooted tree (in Newick format) from a Phylo2Vec matrix}
\usage{
to_newick_from_matrix(input_integers)
}
\description{
Recover a rooted tree (in Newick format) from a Phylo2Vec matrix
}
