% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{to_newick_from_vector}
\alias{to_newick_from_vector}
\title{Recover a rooted tree (in Newick format) from a Phylo2Vec vector}
\usage{
to_newick_from_vector(input_integers)
}
\description{
Recover a rooted tree (in Newick format) from a Phylo2Vec vector
}
