% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{vcov_from_matrix}
\alias{vcov_from_matrix}
\title{Get the variance-covariance matrix of a Phylo2Vec matrix}
\usage{
vcov_from_matrix(matrix)
}
\description{
Get the variance-covariance matrix of a Phylo2Vec matrix
}
