% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{vcv_from_vector}
\alias{vcv_from_vector}
\title{Get the variance-covariance matrix of a Phylo2Vec vector}
\usage{
vcv_from_vector(vector)
}
\description{
Get the variance-covariance matrix of a Phylo2Vec vector
}
