% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R, R/stats.R
\name{cophenetic_distances}
\alias{cophenetic_distances}
\title{Get the (topological) cophenetic distance matrix of a phylo2vec tree}
\usage{
cophenetic_distances(tree, unrooted = FALSE)

cophenetic_distances(tree, unrooted = FALSE)
}
\arguments{
\item{tree}{phylo2vec vector (1D) or matrix (2D)}

\item{unrooted}{If TRUE, compute unrooted distances. Default is FALSE.}
}
\value{
Cophenetic distance matrix (shape: (n_leaves, n_leaves))

Cophenetic distance matrix (shape: (n_leaves, n_leaves))
}
\description{
The cophenetic distance between two leaves is the distance from each leaf to their most recent common ancestor.
For phylo2vec vectors, this is the topological distance. For phylo2vec matrices, this is the distance with branch lengths.

The cophenetic distance between two leaves is the distance from each leaf
to their most recent common ancestor.
For vectors, this is the topological distance.
For matrices, this uses branch lengths.
}
