% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{cophenetic_distances}
\alias{cophenetic_distances}
\title{Compute the cophenetic distance matrix of a precursorhylo2Vec
vector (topological) or matrix (from branch lengths).}
\usage{
cophenetic_distances(vector_or_matrix)
}
\arguments{
\item{n_leaves}{Number of leaves in the tree}

\item{topology_only}{If TRUE, a vector corresponding to the topology
is returned. IF FALSE, a matrix is returned}

\item{ordered}{If TRUE, sample an ordered tree}
}
\value{
A precursorhylo2Vec object (\code{integer} if topology_only, else \code{matrix})
}
\description{
Compute the cophenetic distance matrix of a precursorhylo2Vec
vector (topological) or matrix (from branch lengths).
}
