% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{cophenetic_distances}
\alias{cophenetic_distances}
\title{Get the (topological) cophenetic distance matrix of a phylo2vec tree}
\usage{
cophenetic_distances(tree, unrooted = FALSE)
}
\arguments{
\item{tree}{A phylo2vec tree}

\item{unrooted}{If true, the distance is calculated as the distance between each leaf and their most recent common ancestor, multiplied by 2. If false, the distance is calculated as the distance from each leaf to their most recent common ancestor.}
}
\value{
Cophenetic distance matrix (shape: (n_leaves, n_leaves))
}
\description{
The cophenetic distance between two leaves is the distance from each leaf to their most recent common ancestor.
For phylo2vec vectors, this is the topological distance. For phylo2vec matrices, this is the distance with branch lengths.
}
