% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_newick.R
\name{from_newick}
\alias{from_newick}
\title{Convert a Newick string to a Phylo2Vec vector or matrix}
\usage{
from_newick(newick)
}
\arguments{
\item{newick}{Newick string representation of the tree}
}
\value{
A Phylo2Vec matrix (if branch lengths are present),
otherwise a Phylo2Vec vector (no branch lengths)
}
\description{
Convert a Newick string to a Phylo2Vec vector or matrix
}
