% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_common_ancestor}
\alias{get_common_ancestor}
\title{Get the most recent common ancestor between two nodes in a phylo2vec tree.}
\usage{
get_common_ancestor(vector_or_matrix, node1, node2)
}
\arguments{
\item{vector_or_matrix}{phylo2vec vector (integer) or matrix (numeric)}

\item{node1}{A node in the tree (0-indexed)}

\item{node2}{A node in the tree (0-indexed)}
}
\value{
The most recent common ancestor node (0-indexed)
}
\description{
\code{node1} and \code{node2} can be leaf nodes (0 to n_leaves - 1)
or internal nodes (n_leaves to 2 * n_leaves - 2).
}
\details{
Similar to ape's \code{getMRCA} function (for leaf nodes)
and ETE's \code{get_common_ancestor} (for all nodes), but for phylo2vec.

Note: The MRCA is purely topological, so for matrices, only the vector
component (column 1) is used.
}
