% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_common_ancestor}
\alias{get_common_ancestor}
\title{Get the first recent common ancestor between two nodes in a phylo2vec tree
node1 and node2 can be leaf nodes (0 to n_leaves) or internal nodes (n_leaves to 2*(n_leaves-1)).
Similar to ape's \code{getMRCA} function in R (for leaf nodes)
and ETE's \code{get_common_ancestor} in Python (for all nodes), but for phylo2vec vectors.}
\usage{
get_common_ancestor(vector, node1, node2)
}
\arguments{
\item{vector}{phylo2vec vector representation of a tree topology}

\item{node1}{The first node (0-indexed)}

\item{node2}{The second node (0-indexed)}
}
\value{
The common ancestor node (0-indexed)
}
\description{
Get the first recent common ancestor between two nodes in a phylo2vec tree
node1 and node2 can be leaf nodes (0 to n_leaves) or internal nodes (n_leaves to 2*(n_leaves-1)).
Similar to ape's \code{getMRCA} function in R (for leaf nodes)
and ETE's \code{get_common_ancestor} in Python (for all nodes), but for phylo2vec vectors.
}
