% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_node_depth}
\alias{get_node_depth}
\title{Get the depth of a node in a phylo2vec tree.}
\usage{
get_node_depth(vector_or_matrix, node)
}
\arguments{
\item{vector_or_matrix}{phylo2vec vector (ndim == 1)/matrix (ndim == 2)}

\item{node}{A node in the tree (0-indexed, from 0 to 2*n_leaves - 2)}
}
\value{
Depth of the node (distance from root)
}
\description{
The depth of a node is the length of the path from the root to that node
(i.e., distance from root). This follows the BEAST/ETE convention.
}
\details{
The root has depth 0, and depths increase as you move toward the leaves.

For vectors, topological depth is returned (all branch lengths = 1).
For matrices, actual branch lengths are used.
}
