% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_node_depths}
\alias{get_node_depths}
\title{Get the depths of all nodes in a phylo2vec tree.}
\usage{
get_node_depths(vector_or_matrix)
}
\arguments{
\item{vector_or_matrix}{phylo2vec vector (ndim == 1)/matrix (ndim == 2)}
}
\value{
Vector of depths for all nodes (length = 2 * n_leaves - 1).
Index i+1 contains the depth of node i (R is 1-indexed).
}
\description{
The depth of a node is the length of the path from the root to that node
(i.e., distance from root). This follows the BEAST/ETE convention.
}
\details{
The root has depth 0, and depths increase as you move toward the leaves.

For vectors, topological depth is returned (all branch lengths = 1).
For matrices, actual branch lengths are used.
}
