% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{queue_shuffle}
\alias{queue_shuffle}
\title{Produce an ordered version (i.e., birth-death process version)
of a phylo2vec vector using the Queue Shuffle algorithm.}
\usage{
queue_shuffle(vector, shuffle_cherries)
}
\arguments{
\item{vector}{phylo2vec vector representation of a tree topology}

\item{shuffle_cherries}{If true, the algorithm will randomly shuffle the order of cherries (i.e., pairs of leaves)}
}
\value{
A list with two elements:
\itemize{
\item \code{v}: The ordered phylo2vec vector
\item \code{mapping}: A mapping of the original labels to the new labels
}
}
\description{
Queue Shuffle ensures that the output tree is ordered,
while also ensuring a smooth path through the space of orderings
}
\details{
for more details, see https://doi.org/10.1093/gbe/evad213
}
