% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{remove_branch_lengths}
\alias{remove_branch_lengths}
\title{Remove parent labels from the Newick string}
\usage{
remove_branch_lengths(newick)
}
\description{
Remove parent labels from the Newick string
}
