% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_newick.R
\name{to_newick}
\alias{to_newick}
\title{Convert a phylo2vec vector or matrix to Newick format}
\usage{
to_newick(vector_or_matrix)
}
\arguments{
\item{vector_or_matrix}{phylo2vec matrix (tree with branch
lengths) or a vector (tree topology only)}
}
\value{
Newick string representation of the tree
}
\description{
Convert a phylo2vec vector or matrix to Newick format
}
