% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{to_newick}
\alias{to_newick}
\title{Recover a rooted tree (in Newick format) from a phylo2vec tree}
\usage{
to_newick(tree)
}
\arguments{
\item{tree}{A phylo2vec tree}
}
\value{
Newick string representation of the tree
}
\description{
Recover a rooted tree (in Newick format) from a phylo2vec tree
}
