% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{to_newick}
\alias{to_newick}
\title{Recover a rooted tree (in Newick format) from a Phylo2Vec v}
\usage{
to_newick(input_integers)
}
\description{
Recover a rooted tree (in Newick format) from a Phylo2Vec v
}
