% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{vcovp}
\alias{vcovp}
\title{Compute the variance-covariance matrix of a phylo2vec
vector (tree topology) or matrix (topology + branch lengths).}
\usage{
vcovp(vector_or_matrix)
}
\arguments{
\item{vector_or_matrix}{phylo2vec vector (ndim == 1)/matrix (ndim == 2)}
}
\value{
Variance-covariance matrix
}
\description{
Output shape: (n_leaves, n_leaves)
}
