% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{vcovp}
\alias{vcovp}
\title{Get the variance-covariance matrix of a phylo2vec tree}
\usage{
vcovp(tree)
}
\arguments{
\item{tree}{A phylo2vec tree}
}
\value{
Variance-covariance matrix (shape: (n_leaves, n_leaves))
}
\description{
Get the variance-covariance matrix of a phylo2vec tree
}
